requirements.txt
albumentations==2.0.5
efficientnet_pytorch==0.7.1
einops==0.8.1
ftfy==6.3.1
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
regex==2024.11.6
scikit_learn==1.6.1
seaborn==0.13.2
setuptools==75.8.2
timm==1.0.15
torch==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
transformers==4.49.0
